+++ Changelog.rst
@@ -8,6 +8,8 @@ Changelog
 
 Version compatible with elasticsearch 5.0
 
+ * when using SSL certificate validation is now on by default. Install
+   ``certifi`` or supply root certificate bundle.
  * added ``headers`` arg to connections to support custom http headers
  * passing in a keyword parameter with ``None`` as value will cause that param
    to be ignored